changed all ints to int64

+++ models/category.go
@@ -1,18 +1,18 @@
 package models
 
 type Category struct {
-	ID           *int               `json:"ID,omitempty"`
+	ID           *int64             `json:"ID,omitempty"`
 	DisplayName  *string            `json:"DisplayName,omitempty"`
-	OrderNumber  *int               `json:"OrderNumber,omitempty"`
+	OrderNumber  *int64             `json:"OrderNumber,omitempty"`
 	IsSelected   *bool              `json:"IsSelected,omitempty"`
 	Visible      *bool              `json:"Visible,omitempty"`
-	LanguageID   *int               `json:"LanguageID,omitempty"`
+	LanguageID   *int64             `json:"LanguageID,omitempty"`
 	Tag          *string            `json:"Tag,omitempty"`
 	LanguageCode *string            `json:"LanguageCode,omitempty"`
-	SelectedID   *int               `json:"SelectedID,omitempty"`
+	SelectedID   *int64             `json:"SelectedID,omitempty"`
 	ImageURL     *string            `json:"ImageURL,omitempty"`
-	VendorID     *int               `json:"VendorID,omitempty"`
+	VendorID     *int64             `json:"VendorID,omitempty"`
 	VendorName   *string            `json:"VendorName,omitempty"`
-	ParentId     *int               `json:"ParentId,omitempty"`
+	ParentId     *int64             `json:"ParentId,omitempty"`
 	Devices      *map[string]string `json:"Devices,omitempty"`
 }
