+++ models/countryverificationtypewhitelist.go
@@ -1,7 +1,7 @@
 package models
 
 type CountryVerificationTypeWhitelist struct {
-	CountryID          *int              `json:"CountryID,omitempty"`
-	VerificationTypeID *int              `json:"VerificationTypeID,omitempty"`
+	CountryID          *int64            `json:"CountryID,omitempty"`
+	VerificationTypeID *int64            `json:"VerificationTypeID,omitempty"`
 	VerificationType   *VerificationType `json:"VerificationType,omitempty"`
 }
