+++ models/currenciesobject.go
@@ -1,10 +1,10 @@
 package models
 
 type CurrenciesObject struct {
-	CurrencyId     *int     `json:"CurrencyId,omitempty"`
+	CurrencyId     *int64   `json:"CurrencyId,omitempty"`
 	Code           *string  `json:"Code,omitempty"`
 	Description    *string  `json:"Description,omitempty"`
-	Multiplier     *int     `json:"Multiplier,omitempty"`
+	Multiplier     *int64   `json:"Multiplier,omitempty"`
 	Format         *string  `json:"Format,omitempty"`
 	ExchangeRate   *float64 `json:"ExchangeRate,omitempty"`
 	MaxBetLimit    *float64 `json:"MaxBetLimit,omitempty"`
