+++ models/gamefrontvariant.go
@@ -1,8 +1,8 @@
 package models
 
 type GameFrontVariant struct {
-	GameVariantId     *int     `json:"GameVariantId,omitempty"`
-	GameId            *int     `json:"GameId,omitempty"`
+	GameVariantId     *int64   `json:"GameVariantId,omitempty"`
+	GameId            *int64   `json:"GameId,omitempty"`
 	MinBet            *float64 `json:"MinBet,omitempty"`
 	MaxBet            *float64 `json:"MaxBet,omitempty"`
 	DenominatorAmount *float64 `json:"DenominatorAmount,omitempty"`
