+++ models/paymenttypemodel.go
@@ -1,7 +1,7 @@
 package models
 
 type PaymentTypeModel struct {
-	PaymentTypeID           *int     `json:"PaymentTypeID,omitempty"`
+	PaymentTypeID           *int64   `json:"PaymentTypeID,omitempty"`
 	Name                    *string  `json:"Name,omitempty"`
 	Enabled                 *bool    `json:"Enabled,omitempty"`
 	Created                 *IGCTime `json:"Created,omitempty"`
@@ -20,7 +20,7 @@ type PaymentTypeModel struct {
 	WithdrawFee             *float64 `json:"WithdrawFee,omitempty"`
 	WithdrawMinFee          *float64 `json:"WithdrawMinFee,omitempty"`
 	WithdrawMaxFee          *float64 `json:"WithdrawMaxFee,omitempty"`
-	CaptureAfterMinutes     *int     `json:"CaptureAfterMinutes,omitempty"`
+	CaptureAfterMinutes     *int64   `json:"CaptureAfterMinutes,omitempty"`
 	TwoStepWithdraw         *bool    `json:"TwoStepWithdraw,omitempty"`
 	CardProcessing          *bool    `json:"CardProcessing,omitempty"`
 	BankTransfer            *bool    `json:"BankTransfer,omitempty"`
