+++ models/realitycheckusersetting.go
@@ -1,6 +1,6 @@
 package models
 
 type RealityCheckUserSetting struct {
-	UserId   *int `json:"UserId,omitempty"`   // User identifier
-	Interval *int `json:"Interval,omitempty"` // Interval in seconds
+	UserId   *int64 `json:"UserId,omitempty"`   // User identifier
+	Interval *int64 `json:"Interval,omitempty"` // Interval in seconds
 }
