+++ models/recentwinnersresponsemodel.go
@@ -5,7 +5,7 @@ type RecentWinnersResponseModel struct {
 	CountryCode *string  `json:"CountryCode,omitempty"` // User's registered country code.
 	UserName    *string  `json:"UserName,omitempty"`    // The username.
 	Currency    *string  `json:"Currency,omitempty"`    // The currency of the win.
-	GameID      *int     `json:"GameID,omitempty"`      // The game identifier.
+	GameID      *int64   `json:"GameID,omitempty"`      // The game identifier.
 	GameName    *string  `json:"GameName,omitempty"`    // The name of the game.
 	Win         *float64 `json:"Win,omitempty"`         // The amount of the win.
 	BaseWin     *float64 `json:"BaseWin,omitempty"`     // The base currency amount of the win.
