+++ models/rolesresponse.go
@@ -1,6 +1,6 @@
 package models
 
 type RolesResponse struct {
-	Id   *int    `json:"Id,omitempty"`
+	Id   *int64  `json:"Id,omitempty"`
 	Name *string `json:"Name,omitempty"`
 }
