+++ models/selfexclusioncategoryduration.go
@@ -1,7 +1,7 @@
 package models
 
 type SelfExclusionCategoryDuration struct {
-	ID     *int `json:"ID,omitempty"`
-	Days   *int `json:"Days,omitempty"`   // Duration in day units.
-	Months *int `json:"Months,omitempty"` // Duration in month units.
+	ID     *int64 `json:"ID,omitempty"`
+	Days   *int64 `json:"Days,omitempty"`   // Duration in day units.
+	Months *int64 `json:"Months,omitempty"` // Duration in month units.
 }
